+++ src/billiards/simulation.py
@@ -4,16 +4,18 @@ from math import isinf
 
 import numpy as np
 
+from .obstacles import Obstacle
 from .physics import elastic_collision, time_of_impact
 
 INF = float("inf")
 
 
 class Billiard(object):
-    """The billiard class represent a billiard world.
+    """The billiard class represent a billiard table.
 
-    By default the 2-dimensional world is infinite in every direction.
-    To populate it use the `add_ball` method.
+    By default the 2-dimensional table is infinite in every direction, but you
+    can place static obstacles on them. use the `add_ball` method to add some
+    billiard balls.
 
     Attributes:
         time: Current simulation time.
@@ -25,11 +27,26 @@ class Billiard(object):
         toi_table: Lower-triangular matrix (= nested lists) of time of impacts.
         toi_min: List of time-index pairs of the next collision for each ball.
         toi_next: Time-index-index triple for the next collision.
+        obstacles: List of obstacles on the table.
+        obstacles_toi: time-obstacle pairs for each ball.
+        obstacles_next: Time-index-obstacle pair of the next obstacle impact.
 
     """
 
-    def __init__(self):
-        """Create an empty world."""
+    def __init__(self, obstacles=None):
+        """Setup a billiard table with the given obstacles but without balls.
+
+        Args:
+            obstacles: Iterable containing billiards.obstacle.Obstacle objects.
+
+        Raises:
+            TypeError: If one of the obstacles is not a
+            billiards.obstacle.Obstacle instance.
+        """
+        if obstacles is None:
+            obstacles = []
+
+        # initalize time and number of balls to 0
         self.time = 0.0
         self.num = 0
 
@@ -39,10 +56,23 @@ class Billiard(object):
         self.balls_radius = []
         self.balls_mass = []
 
-        # time of impact table
+        # time of impact table for ball-ball collisions
         self.toi_table = []
         self.toi_min = []
         self.toi_next = (INF, -1, 0)
+
+        # check obstacles
+        self.obstacles = []
+        for obs in obstacles:
+            if not isinstance(obs, Obstacle):
+                msg = "{} must be an obstacle.Obstacle instance"
+                raise TypeError(msg.format(obs))
+
+            self.obstacles.append(obs)
+
+        # time of impact list for ball-obstacle collisions
+        self.obstacles_toi = []  # next time of impact for each ball
+        self.obstacles_next = (INF, -1, None)  # next obstacle collision
 
     def add_ball(self, pos, vel, radius=0.0, mass=1.0):
         """Add a ball at the given position with the given velocity.
@@ -88,12 +118,19 @@ class Billiard(object):
 
         self.toi_next = min((t, j, i) for i, (t, j) in enumerate(self.toi_min))
 
+        # calculate time of impact for obstacles
+        self.obstacles_toi.append(self.calc_next_obstacle(idx))
+        self.obstacles_next = min(
+            (t, i, obs) for i, (t, obs) in enumerate(self.obstacles_toi)
+        )
+
         # sanity checks
         assert self.balls_position.shape[0] == self.num
         assert self.balls_velocity.shape[0] == self.num
         assert len(self.balls_radius) == self.num
         assert len(self.balls_mass) == self.num
         assert len(self.toi_table) == self.num
+        assert len(self.obstacles_toi) == self.num
 
         return idx
 
@@ -117,6 +154,28 @@ class Billiard(object):
         r2 = self.balls_radius[idx2]
 
         return self.time + time_of_impact(p1, v1, r1, p2, v2, r2)
+
+    def calc_next_obstacle(self, idx):
+        """Find the closest colliding obstacle for the given ball.
+
+        Args:
+            idx: Index of ball.
+
+        Returns:
+            tuple: (time, obstacle)-pair of the next collision or (INF, None)
+            if ball will not impact any obstacle.
+        """
+        pos = self.balls_position[idx]
+        vel = self.balls_velocity[idx]
+        radius = self.balls_radius[idx]
+
+        t_min, obs_min = INF, None
+        for obs in self.obstacles:
+            t = obs.calc_toi(pos, vel, radius)
+            if t < t_min:
+                t_min, obs_min = t, obs
+
+        return self.time + t_min, obs_min
 
     def collide_balls(self, idx1, idx2):
         """Update the velocities of two colliding balls in the simulation.
@@ -144,31 +203,69 @@ class Billiard(object):
         self.balls_velocity[idx1] = v1
         self.balls_velocity[idx2] = v2
 
+    def collide_obstacle(self, idx, obs):
+        """Update velocity of a ball colliding with an obstacle.
+
+        Args:
+            idx: Index of the colliding ball.
+            obs: The obstacle with which the ball collides.
+
+        """
+        pos = self.balls_position[idx]
+        vel = self.balls_velocity[idx]
+        radius = self.balls_radius[idx]
+
+        new_vel = obs.collide(pos, vel, radius)
+        self.balls_velocity[idx] = new_vel
+
     def evolve(self, end_time):
         """Advance the simulation until the given time is reached.
+
+        Will call bounce_ballball and bounce_ballobstacle repeatetly (which one
+        depends on self.toi_next and self.obstacles_next) until the end_time
+        is reached.
 
         Args:
             end_time: Time until which the billiard should be simulated.
 
-        """
-        while self.toi_next[0] <= end_time:
-            self.bounce()
+        Returns:
+            list: List of collisions, each item is a (float, int, int).triplet
+            where the first number is the time of impact and the next two
+            integers are the indices of the balls that collided.
 
-        assert end_time < self.toi_next[0]
+        """
+        collisions = []
+        while min(self.toi_next[0], self.obstacles_next[0]) <= end_time:
+            # decide what kind of collision we are dealing with
+            if self.toi_next[0] <= self.obstacles_next[0]:
+                coll = self.bounce_ballball()
+            else:
+                coll = self.bounce_ballobstacle()
+
+            collisions.append(coll)
+
+        assert end_time < min(self.toi_next[0], self.obstacles_next[0])
         self._move(end_time)
 
-    def bounce(self):
-        """Advance to the next collision and handle it."""
-        t, idx1, idx2 = self.toi_next
+        return collisions
 
-        # advance to the next collision
-        self._move(t)
+    def bounce_ballball(self):
+        """Advance to the next ball-ball collision and handle it.
+
+        Returns:
+            float: Time of impact of the just handled collision.
+            int: Index of one ball that collided.
+            int: Index of the other ball.
+
+        """
+        t, idx1, idx2 = self.toi_next
 
         # get the balls that collide
         assert idx1 < idx2, (idx1, idx2)
         assert self.toi_min[idx2] == (t, idx1)
 
-        # handle collision between balls with index idx1 and idx2
+        # advance to the next collision and handle it
+        self._move(t)
         self.collide_balls(idx1, idx2)
 
         # update time of impact for the two balls
@@ -195,6 +292,57 @@ class Billiard(object):
             self.toi_min[i] = toi_idx
 
         self.toi_next = min((t, i, j) for j, (t, i) in enumerate(self.toi_min))
+
+        # update time of impact for the next ball-obstacle collision
+        self.obstacles_toi[idx1] = self.calc_next_obstacle(idx1)
+        self.obstacles_toi[idx2] = self.calc_next_obstacle(idx2)
+
+        self.obstacles_next = min(
+            (t, i, obs) for i, (t, obs) in enumerate(self.obstacles_toi)
+        )
+
+        return (t, idx1, idx2)
+
+    def bounce_ballobstacle(self):
+        """Advance to the next ball-obstacle collision and handle it.
+
+        Returns:
+            float: Time of impact of the just handled collision.
+            int: Index of the ball that collided.
+            billiards.obstacle.Obstacle: Obstacle that the ball collided with.
+
+        """
+        t, idx, obs = self.obstacles_next
+        assert self.obstacles_toi[idx] == (t, obs)
+
+        # advance to the next collision and handle it
+        self._move(t)
+        self.collide_obstacle(idx, obs)
+
+        # update time of impact for ball-ball collisions
+        for j in range(idx):
+            self.toi_table[idx][j] = self.calc_toi(idx, j)
+
+        for i in range(idx + 1, self.num):
+            self.toi_table[i][idx] = self.calc_toi(i, idx)
+
+        # update toi_min
+        for i in range(idx, self.num):
+            row = self.toi_table[i]
+            toi_pairs = ((t, j) for (j, t) in enumerate(row))
+            toi_idx = min(toi_pairs, default=(INF, -1))
+            self.toi_min[i] = toi_idx
+
+        self.toi_next = min((t, i, j) for j, (t, i) in enumerate(self.toi_min))
+
+        # update time of impact for the next ball-obstacle collision
+        self.obstacles_toi[idx] = self.calc_next_obstacle(idx)
+
+        self.obstacles_next = min(
+            (t, i, obs) for i, (t, obs) in enumerate(self.obstacles_toi)
+        )
+
+        return (t, idx, obs)
 
     def _move(self, time):
         # just update position, no collision handling here